+++ src/main/java/uk/co/duelmonster/minersadvantage/client/ModMenuEntrypoint.java
@@ -7,7 +7,7 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
 public final class ModMenuEntrypoint implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return parent -> null;
+        return MinersAdvantageConfigScreen::create;
     }
 }
 //?} else {